two services and two controllers done

+++ movie-app-task-backend/Models/Rating.cs
@@ -8,7 +8,7 @@ namespace movie_app_task_backend.Models
     {
         public int Id { get; set; }
 
-        public int Rating_value { get; set; }
+        public float Rating_value { get; set; }
 
         public int MediaId { get; set; }
